New image: pipeline with rstudio

+++ start.sh
@@ -1,4 +1,4 @@
-#!/bin/sh
+#!/bin/bash
 # This is a comment!
 echo Hello World
 
@@ -12,15 +12,19 @@ mkdir -p $HEALTHYDIR
 LOGSDIR="data/logs"
 mkdir -p $LOGSDIR
 
-gdc-client download -d $CANCERDIR -m manifests/${PRIMARY_SITE}.cancer.short.txt --log-file ${LOGSDIR}/${PRIMARY_SITE}.cancer.log --retry-amount 3
-gdc-client download -d $HEALTHYDIR -m manifests/${PRIMARY_SITE}.healthy.short.txt --log-file ${LOGSDIR}/${PRIMARY_SITE}.healthy.log --retry-amount 3
+if [[ $DOWNLOAD == "true" ]]; then
+  gdc-client download -d $CANCERDIR -m manifests/${PRIMARY_SITE}.cancer.short.txt --log-file ${LOGSDIR}/${PRIMARY_SITE}.cancer.log --retry-amount 3
+  gdc-client download -d $HEALTHYDIR -m manifests/${PRIMARY_SITE}.healthy.short.txt --log-file ${LOGSDIR}/${PRIMARY_SITE}.healthy.log --retry-amount 3
 
-find $CANCERDIR -name '*.gz' -exec mv '{}' $CANCERDIR \;
-find $HEALTHYDIR -name '*.gz' -exec mv '{}' $HEALTHYDIR \;
+  find $CANCERDIR -name '*.gz' -exec mv '{}' $CANCERDIR \;
+  find $HEALTHYDIR -name '*.gz' -exec mv '{}' $HEALTHYDIR \;
 
-find . -type d -empty -delete
-find . -name '*.gz' -exec gunzip '{}' \;
-find . -name '*.gz' -exec rm '{}' \;
+  find . -type d -empty -delete
+  find . -name '*.gz' -exec gunzip '{}' \;
+  find . -name '*.gz' -exec rm '{}' \;
+
+echo Finished downloading
+fi
 
 Rscript /01_GetTheData.R $DATADIR
 Rscript /02_QC.R $DATADIR
